refactor(lib): organize utility files into domain-specific subfolders
This commit refactors the structure of the `/lib` directories to improve organization, clarity, and scalability of the project's core utilities.

- **API Helpers:** Grouped data-fetching utilities (`fetcher.js`, `fetchServerData.js`) into `lib/api`.
- **SEO:** Grouped SEO helpers (`metadata.js`) into `lib/seo`.
- **General Utilities:** Placed general-purpose utilities (`logger.js`, `errorHandler.js`, `errors.js`,`db.js`) into `lib/utils`.
- **Providers:** Created a new `providers` folder for any upcoming client-side React Context providers.
